+++ src/Backup.ts
@@ -22,6 +22,7 @@ class Backup {
   private zipArchive: string;
   private compressionLevel: number;
   private singleJex: boolean;
+  private createSubfolder: boolean;
   private backupSetName: string;
 
   constructor() {
@@ -41,15 +42,17 @@ class Backup {
     await this.createErrorDialog();
     await this.loadSettings();
     await this.startTimer();
-    await this.upgradeBackupTargetVersion();
+    await this.upgradeBackupPluginVersion();
     await sevenZip.updateBinPath();
     await sevenZip.setExecutionFlag();
     this.backupStartTime = null;
   }
 
-  private async upgradeBackupTargetVersion() {
-    let version = await joplin.settings.value("backupVersion");
-    const targetVersion = 1;
+  private async upgradeBackupPluginVersion() {
+    this.log.verbose("Upgrade Backup Plugin");
+    let startVersion = await joplin.settings.value("backupVersion");
+    let version = startVersion;
+    const targetVersion = 2;
     for (
       let checkVersion = version + 1;
       checkVersion <= targetVersion;
@@ -59,6 +62,17 @@ class Backup {
         if (checkVersion === 1) {
           if (this.backupBasePath !== "" && this.backupRetention > 1) {
             await this.saveOldBackupInfo();
+          }
+        } else if (checkVersion === 2) {
+          // When a path is set from old installation, disable the subfolder creation
+          if (
+            this.backupBasePath &&
+            this.backupBasePath !== "" &&
+            startVersion === 1
+          ) {
+            this.log.verbose("createSubfolder: false");
+            this.createSubfolder = false;
+            await joplin.settings.setValue("createSubfolder", false);
           }
         }
 
@@ -202,6 +216,20 @@ class Backup {
       );
     }
 
+    if (this.createSubfolder) {
+      this.log.verbose("append subFolder");
+      this.backupBasePath = path.join(this.backupBasePath, "JoplinBackup");
+      if (!fs.existsSync(this.backupBasePath)) {
+        try {
+          fs.mkdirSync(this.backupBasePath);
+        } catch (e) {
+          this.backupBasePath = null;
+          await this.showError("create Folder: " + e.message);
+          throw e;
+        }
+      }
+    }
+
     if (path.normalize(profileDir) === this.backupBasePath) {
       this.backupBasePath = null;
     }
@@ -210,6 +238,7 @@ class Backup {
 
   public async loadSettings() {
     this.log.verbose("loadSettings");
+    this.createSubfolder = await joplin.settings.value("createSubfolder");
     await this.loadBackupPath();
     this.backupRetention = await joplin.settings.value("backupRetention");